requirement.txt
pipenv
python-dotenv
spotipy
pandas
pillow
urllib